Refactored database structure for clarity and consistency
This commit restructures and refines the naming schemes and functionalities of the database used to store details related to modules, environments and file registries.

The tables were renamed from camelCase to lower_snake_case to adhere to SQL naming conventions and improve readability. Columns such as 'created_at' and 'updated_at' were modified to specifically represent the elements they correspond to (like 'module_created_at').

These changes will bring enhanced consistency across the database structure and increase its intuitiveness, aiding in easier management of the data. The corresponding SQL queries in the repositories and the assertions in the test cases were updated to reflect the changes.

The change also involved modifying methods that were previously named in Portuguese to use English for maintaining code uniformity and clarity.

Updating the naming schemes in the database also involved modifying the respective repositories and test classes to ensure that the correct column names are referenced.

Further, the 'executeSQL' method is refactored to accurately describe the 'Statement' object as a 'statement' as it more accurately conveys its functionality.

The changes might not impact the immediate functionality of the application, but it significantly augments code readability and maintainability.